+++ ftplugin/cabal.lua
@@ -1,2 +1,3 @@
 local ht = require('haskell-tools.internal')
 ht.start_or_attach()
+ht.dap_discover()